requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alembic=1.12.0=pypi_0
altair=4.2.2=pypi_0
anyio=4.0.0=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
avro=1.11.0=pypi_0
babel=2.13.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
blinker=1.6.3=pypi_0
boto3=1.28.65=pypi_0
botocore=1.31.65=pypi_0
bottleneck=1.3.7=py39h7b7cdfe_0
brotli-python=1.0.9=py39h7a8716b_9
build=1.0.3=pypi_0
ca-certificates=2024.3.11=hecd8cb5_0
cachecontrol=0.14.0=pyhd8ed1ab_0
cachecontrol-with-filecache=0.14.0=pyhd8ed1ab_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=py39h6c40b1e_0
charset-normalizer=3.3.0=pypi_0
cleo=2.1.0=py39hecd8cb5_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
colorama=0.4.6=py39hecd8cb5_0
colorlog=6.7.0=pypi_0
comet-ml=3.35.3=pypi_0
comm=0.1.4=pypi_0
configobj=5.0.8=pypi_0
confluent-kafka=1.9.0=pypi_0
crashtest=0.4.1=py39hecd8cb5_0
cryptography=41.0.4=pypi_0
dataclasses=0.6=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
discordwebhook=1.0.3=pypi_0
distlib=0.3.8=pyhd8ed1ab_0
docopt=0.6.2=py39hecd8cb5_1
dulwich=0.21.6=pypi_0
entrypoints=0.4=pypi_0
everett=3.1.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.0=pypi_0
fastavro=1.8.2=pypi_0
fastjsonschema=2.18.1=pypi_0
filelock=3.13.1=py39hecd8cb5_0
fiona=1.9.5=pypi_0
fire=0.5.0=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2023.9.2=pypi_0
furl=2.1.3=pypi_0
future=0.18.3=pypi_0
geopandas=0.14.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.38=pypi_0
great-expectations=0.14.13=pypi_0
greenlet=3.0.0=pypi_0
h11=0.14.0=pypi_0
hopsworks=3.4.2=pypi_0
hsfs=3.4.2=pypi_0
hsml=3.4.2=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
intel-openmp=2023.1.0=ha357a0b_43548
ipykernel=6.25.2=pypi_0
ipython=8.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jaraco.classes=3.2.1=pyhd3eb1b0_0
javaobj-py3=0.4.3=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.0.3=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pyhd8ed1ab_0
json5=0.9.14=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.4.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.4.0=pypi_0
jupyter-events=0.8.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.8.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.7=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keyring=24.3.1=py39hecd8cb5_0
krb5=1.20.1=h428f121_1
libcxx=14.0.6=h9765a3e_0
libedit=3.1.20230828=h6c40b1e_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libopenblas=0.3.21=h54e7dc3_0
libsodium=1.0.18=hbcb3906_1
lightgbm=4.1.0=py39hcec6c5f_0
llvm-openmp=18.1.2=hb6ac08f_0
mako=1.2.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mock=5.1.0=pypi_0
more-itertools=10.1.0=py39hecd8cb5_0
msgpack-python=1.0.3=py39haf03e11_0
nbclient=0.8.0=pypi_0
nbconvert=7.9.2=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.8=pypi_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
numexpr=2.8.7=py39hf6dca73_0
numpy=1.26.1=pypi_0
openssl=3.0.13=hca72f7f_0
optuna=3.4.0=pypi_0
orderedmultidict=1.0.1=pypi_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3=pypi_0
pip-tools=6.14.0=pypi_0
pkginfo=1.9.6=py39hecd8cb5_0
platformdirs=3.11.0=pypi_0
plotly=5.17.0=pypi_0
pluggy=0.13.1=pypi_0
poetry=1.8.2=osx_pyh534df25_0
poetry-core=1.9.0=pyhd8ed1ab_0
poetry-plugin-export=1.7.1=pyhd8ed1ab_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=4.24.4=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd8ed1ab_0
py=1.11.0=pypi_0
pyarrow=10.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.19.0=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.16.1=pypi_0
pyhopshive=0.6.4.1.dev0=pypi_0
pyhumps=1.6.1=pypi_0
pyjks=20.0.0=pypi_0
pykerberos=1.2.1=py39h6c40b1e_4
pymysql=1.1.0=pypi_0
pyparsing=2.4.7=pypi_0
pyproj=3.6.1=pypi_0
pyproject_hooks=1.0.0=py39hecd8cb5_0
pysocks=1.7.1=pyha2e5f31_6
pytest=5.4.3=pypi_0
python=3.9.19=h5ee71fb_0
python-box=6.1.0=pypi_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-hdfs=2.7.3=pyhd8ed1ab_0
python-installer=0.7.0=py39hecd8cb5_0
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
rapidfuzz=3.5.2=py39hcec6c5f_0
readline=8.2=hca72f7f_0
referencing=0.30.2=pypi_0
requests=2.31.0=py39hecd8cb5_1
requests-kerberos=0.12.0=py39hecd8cb5_0
requests-toolbelt=1.0.0=py39hecd8cb5_0
retry=0.9.2=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.6.0=pypi_0
rpds-py=0.10.6=pypi_0
ruamel-yaml=0.17.17=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
s3transfer=0.7.0=pypi_0
scikit-learn=1.3.1=pypi_0
scipy=1.11.3=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.2=pypi_0
sentry-sdk=1.35.0=pypi_0
setuptools=68.2.2=py39hecd8cb5_0
shapely=2.0.2=pypi_0
shellingham=1.5.0=py39hecd8cb5_0
simplejson=3.19.2=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.22=pypi_0
sqlite=3.41.2=h6c40b1e_0
src=0.1.0=pypi_0
stack-data=0.6.3=pypi_0
streamlit=1.27.2=pypi_0
tbb=2021.8.0=ha357a0b_0
tenacity=8.2.3=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pypi_0
thrift=0.16.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
tomli=2.0.1=py39hecd8cb5_0
tomlkit=0.12.4=pyha770c72_0
toolz=0.12.0=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.11.2=pypi_0
trove-classifiers=2023.10.18=py39hecd8cb5_0
twofish=0.3.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.18=pypi_0
validators=0.22.0=pypi_0
virtualenv=20.25.1=pyhd8ed1ab_0
watchdog=2.3.1=pypi_0
wcwidth=0.2.8=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.3=pypi_0
wheel=0.41.2=py39hecd8cb5_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.16.0=pypi_0
wurlitzer=3.0.3=pypi_0
xattr=1.1.0=py39ha09f3b3_0
xgboost=2.0.0=pypi_0
xz=5.4.6=h6c40b1e_0
zeromq=4.3.5=hcec6c5f_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h4dc903c_0
